hotfix/ clean initial files
