+++ src/App.jsx
@@ -1,0 +1,17 @@
+import './App.css';
+import InvoiceAssigner from './views/InvoiceAssigner';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>Bemmbo</h1>
+      </header>
+      <div className="App-body">
+        <InvoiceAssigner/>
+      </div>
+    </div>
+  );
+}
+
+export default App;
